+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## Unreleased
+
+### Fixes
+
+*  auth: make emeis user model able to authenticate (#249) 
+
 ## v0.2.0 (22 July 2021)
 
 Various bugfixes to make Emeis usable as a Django app